Compute confidence intervals for sensitivity indices

Computes confidence intervals for a gsaot_indices object using bootstrap results.

Usage

# S3 method for class 'gsaot_indices'
confint(object, parm = NULL, level = 0.95, type = "norm", ...)

Arguments

object

An object of class gsaot_indices, with bootstrap results included.

parm

A specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

(default is 0.95) Confidence level for the interval.

type

(default is "norm") Method to compute the confidence interval. For more information, check the type option of boot::boot.ci().

...

Additional arguments (currently unused).

Value

A data frame with the following columns:

input: Name of the input variable.

component: The index component for Wasserstein-Bures.

index: Estimated indices

original: Original estimates.

bias: Bootstrap bias estimate.

low.ci: Lower bound of the confidence interval.

high.ci: Upper bound of the confidence interval.

Examples

N <- 1000

mx <- c(1, 1, 1)
Sigmax <- matrix(data = c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3)

x1 <- rnorm(N)
x2 <- rnorm(N)
x3 <- rnorm(N)

x <- cbind(x1, x2, x3)
x <- mx + x %*% chol(Sigmax)

A <- matrix(data = c(4, -2, 1, 2, 5, -1), nrow = 2, byrow = TRUE)
y <- t(A %*% t(x))

x <- data.frame(x)
y <- y

res <- ot_indices_wb(x, y, 10, boot = TRUE, R = 1000)
confint(res, parm = c(1,3), level = 0.9)
#>   input  component      index   original        bias      low.ci   high.ci
#> 1    X1 wass-bures 0.46835173 0.47235449 0.004002761 0.453210817 0.4834926
#> 2    X3 wass-bures 0.12634157 0.13312687 0.006785299 0.109245611 0.1434375
#> 3    X1  advective 0.29501187 0.29692734 0.001915474 0.285555655 0.3044681
#> 4    X3  advective 0.11566000 0.11918158 0.003521584 0.101724230 0.1295958
#> 5    X1  diffusive 0.17333986 0.17542715 0.002087287 0.166513995 0.1801657
#> 6    X3  diffusive 0.01068157 0.01394529 0.003263715 0.007006048 0.0143571