Ishigami-Homma function evaluation

Evaluates the Ishigami-Homma function. Input samples are drawn from a uniform distribution over \([-\pi, \pi]^3\)

Usage

ishi_homma_fun(N, A = 2, B = 1)

Arguments

N

Number of input samples to generate.

A

(default: 2) Numeric, amplitude of the second sine component .

B

(default: 1) Numeric, coefficient of the interaction term.

Value

A list with two elements:

x: a numeric matrix of size N x 8 containing the input samples.

y: a numeric vector of length N with the corresponding function outputs.

Details

The Ishigami-Homma function is defined as: $$Y = \sin(X_1) + A \cdot \sin^2(X_2) + B \cdot X_3^4 \cdot \sin(X_1)$$ where \(X_i \sim \mathcal{U}(-\pi, \pi)\).

See also

Examples

result <- ishi_homma_fun(1000)
#> Warning: data length [8000] is not a sub-multiple or multiple of the number of rows [2667]
head(result$x)
#>              X1         X2         X3
#> [1,]  1.0339420  2.1994534 -2.9752946
#> [2,]  1.0053054 -0.7234555 -0.9522839
#> [3,] -0.6044416  1.6351862 -1.1976354
#> [4,]  2.9443150  1.5296528  2.9740846
#> [5,] -0.3675223 -1.9738916  2.6821323
#> [6,]  1.0723878  2.1302988 -3.0769914
head(result$y)
#> [1]  69.5080920   2.4151027   0.2542454  17.5271510 -17.2614975  81.0503157