Sobol G function evaluation

This function evaluates the Sobol G function on a set of input samples generated via crude Monte Carlos. It returns both the sampled inputs and the corresponding function outputs.

Usage

sobol_fun(N, a = c(0, 1, 4.5, 9, 99, 99, 99, 99))

Arguments

N

Integer. Number of input samples to generate.

a

(default: c(0, 1, 4.5, 9, 99, 99, 99, 99)) Numeric vector of non-negative parameters of length 8. These parameters control the sensitivity of each input dimension.

Value

A list with two elements:

x: a numeric matrix of size N x 8 containing the input samples.

y: a numeric vector of length N with the corresponding function outputs.

Details

The Sobol G function is defined as: $$ Y = \prod_{j=1}^{8} \frac{|\ 4 X_j - 2\ | + a_j}{1 + a_j} $$ where \(X_j \sim \mathcal{U}(0, 1)\) independently.

See also

Examples

result <- sobol_fun(1000)
head(result$x)
#>               X1         X2         X3         X4         X5        X6
#> [1,] 0.819940696 0.71294610 0.08217868 0.08814804 0.27732436 0.2969510
#> [2,] 0.005341916 0.17030615 0.71545299 0.10794503 0.89316649 0.7459970
#> [3,] 0.933079347 0.57950210 0.25247931 0.88603579 0.55434687 0.4848396
#> [4,] 0.034009106 0.44031980 0.06267785 0.29595151 0.33588903 0.1310767
#> [5,] 0.572111856 0.23469537 0.22625732 0.99652996 0.03911218 0.6517347
#> [6,] 0.532004169 0.08975488 0.01399731 0.41825278 0.22266917 0.1347008
#>             X7        X8
#> [1,] 0.8859099 0.7896342
#> [2,] 0.2066326 0.6733328
#> [3,] 0.8208537 0.1227273
#> [4,] 0.4112295 0.4428259
#> [5,] 0.4395499 0.5921693
#> [6,] 0.7927640 0.5015832
head(result$y)
#> [1] 1.4213400 2.3734154 1.1903277 1.2711173 0.3290956 0.1842531